refactor: move package under src/
