+++ tests/resources/test_segment.py
@@ -1,6 +1,5 @@
-from exceptions import NotFoundError
-
-from metabase.resources.segment import Segment
+from metabase import Segment
+from metabase.exceptions import NotFoundError
 from tests.helpers import IntegrationTestCase
